requirements.txt
pytest
playwright
allure-pytest
requests
axe-playwright-python==0.1.4
